fix(api): [#143] do not fail trying to remove a whitelisted torrent twice
Previous behavior:

When you try to remove a non exisinting whitelisted torrent the response
is 500 error.

New bahavior:

The enpoints checks if the torrent is included in the whitelist. If it
does not, then it ignores the reqeust returning a 200 code.

It should return a 204 or 404 but the current API only uses these codes:
200, 400, 405, 500. In the new API version we are planning to refctor
all endpoints.

+++ src/tracker/mod.rs
@@ -152,9 +152,28 @@ impl Tracker {
     ///
     /// Will return a `database::Error` if unable to remove the `info_hash` from the whitelist database.
     pub async fn remove_torrent_from_whitelist(&self, info_hash: &InfoHash) -> Result<(), databases::error::Error> {
-        self.database.remove_info_hash_from_whitelist(*info_hash).await?;
-        self.whitelist.write().await.remove(info_hash);
+        self.remove_torrent_from_database_whitelist(info_hash).await?;
+        self.remove_torrent_from_memory_whitelist(info_hash).await;
         Ok(())
+    }
+
+    /// # Errors
+    ///
+    /// Will return a `database::Error` if unable to remove the `info_hash` from the whitelist database.
+    pub async fn remove_torrent_from_database_whitelist(&self, info_hash: &InfoHash) -> Result<(), databases::error::Error> {
+        let is_whitelisted = self.database.is_info_hash_whitelisted(info_hash).await?;
+
+        if !is_whitelisted {
+            return Ok(());
+        }
+
+        self.database.remove_info_hash_from_whitelist(*info_hash).await?;
+
+        Ok(())
+    }
+
+    pub async fn remove_torrent_from_memory_whitelist(&self, info_hash: &InfoHash) -> bool {
+        self.whitelist.write().await.remove(info_hash)
     }
 
     pub async fn is_info_hash_whitelisted(&self, info_hash: &InfoHash) -> bool {
